requirements.txt
aiohttp==3.6.2
alpha-vantage==2.2.0
appnope==0.1.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.4
blpapi==3.14.0
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-bootstrap3==12.0.3
django-braces==1.14.0
django-filter==2.2.0
django-import-export==2.0.2
django-storages==1.9.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
forex-python==1.5
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
inflection==0.4.0
investpy==0.9.14
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.2
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.2.0
lxml==4.5.0
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.1
misaka==2.1.1
mistune==0.8.4
more-itertools==8.2.0
mpld3==0.3
multidict==4.7.5
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
nsepy==0.8
numpy==1.18.3
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pdblp==0.1.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
Quandl==3.5.0
requests==2.23.0
s3transfer==0.3.3
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tablib==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.4.2
yfinance==0.1.54
zipp==2.2.0
